Fill in the Rudder Spec and Status, in api/v1alpha1/rudder_types.go. Take note of the kubebuilder validation directives.
Fill in the Rudder controller in controllers/rudder_controller.go.
Take note of the comment regarding the use of WithName() for logging.

To make it easier to work with a KIND environment, adjust the IMG value in
the Makefile and add a kind-push target.

After filling in the rudder_types.go file, run "make manifests" and
"make generate".  The "make manifests" will create a new OpenAPI CRD and other
supporting pieces for the Rudder in the config/ dir.  If our RudderSpec
and RudderStatus types had been more complex then then "make generate" would
have updated the deepcopy routines in api/v1alpha1/zz_generated.deepcopy.go,
but in this case that was not necessary.

Add "bin/" to .gitignore.  The "make manifests" and "make generate" targets
can add things to bin/, and we don't want to commit those.

Add three example Rudder resources in config/samples.  The comments in these
files describe some nuances and issues with the use of "omitempty" in "type
Rudder struct" in api/v1alpha1/rudder_types.go, regarding whether the
kubebuilder:default directives in the RudderSpec and RudderStatus will be
executed.