Fill in the airplane CRD, adding a tail number for our airplane.
Make the airplane controller build and hook up a Peddles resource and a
Rudder resource.

Hook up the Airplane controller to the test suite.

Add a test for the airplane controller.

Run "make manifests" to create a new OpenAPI CRD manifest.

Run "make generate" to rebuild the deepcopy library for the complex
airplane data type.